# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Template API.",
        "title": "Go - Template API",
        "contact": {
            "name": "Raul del Aguila",
            "email": "email@email.com"
        }
    },
    "basePath": "/",
    "paths": {
...
        "/operator": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get operators",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Get operators",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language responses",
                        "name": "lang",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "descending order 'desc' or ascending order 'asc'",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "'updated_at', 'created_at', 'name' or some other field of the response object",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ItemsOutputDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httphelper.HTTPResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Insert operator",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Insert operator",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language responses",
                        "name": "lang",
                        "in": "query"
                    },
                    {
                        "description": "Operator model",
                        "name": "operator",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OperatorInputDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Operator"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httphelper.HTTPResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/httphelper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httphelper.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/operator/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get operator by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Get operator by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language responses",
                        "name": "lang",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Operator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Operator"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httphelper.HTTPResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httphelper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httphelper.HTTPResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update operator by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Update operator by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language responses",
                        "name": "lang",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Operator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Operator model",
                        "name": "operator",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OperatorInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Operator"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httphelper.HTTPResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httphelper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httphelper.HTTPResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete operator by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Delete operator by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language responses",
                        "name": "lang",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Operator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httphelper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httphelper.HTTPResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Operator": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "minLength": 2
                }
            }
        },
...
        "dto.ItemsOutputDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {}
            }
        },
...
        "dto.OperatorInputDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "TIM"
                }
            }
        },
...
        "httphelper.HTTPResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}